address comments

+++ pkg/tsoutil/tso.go
@@ -21,9 +21,9 @@ const (
 )
 
 // ParseTS parses the ts to (physical,logical).
-func ParseTS(ts int64) (time.Time, int64) {
+func ParseTS(ts uint64) (time.Time, uint64) {
 	logical := ts & logicalBits
 	physical := ts >> physicalShiftBits
-	physicalTime := time.Unix(physical/1000, int64(physical%1000)*time.Millisecond.Nanoseconds())
+	physicalTime := time.Unix(int64(physical/1000), int64(physical)%1000*time.Millisecond.Nanoseconds())
 	return physicalTime, logical
 }
